Make liveness probe optional
